refactor(app): Rename sync_position_to_dlc
